Chore: Add setup process for artifact creation, also move proving params to better location
