+++ proofs/src/lib.rs
@@ -23,6 +23,7 @@ pub mod circom;
 pub mod errors;
 pub mod program;
 pub mod proof;
+pub mod setup;
 #[cfg(test)] pub(crate) mod tests;
 pub mod witness;
 
@@ -47,71 +48,3 @@ pub type ProverKey = client_side_prover::supernova::snark::ProverKey<E1, S1, S2>
 /// The `VerifierKey` needed to create a `CompressedSNARK` using the `proofs` crate choices of
 /// curves and engines.
 pub type VerifierKey = client_side_prover::supernova::snark::VerifierKey<E1, S1, S2>;
-
-use std::io::Cursor;
-
-use client_side_prover::fast_serde::{SerdeByteError, SerdeByteTypes};
-
-#[derive(Debug)]
-pub struct ProvingParams {
-  pub aux_params:          AuxParams,
-  pub vk_digest_primary:   <E1 as Engine>::Scalar,
-  pub vk_digest_secondary: <Dual<E1> as Engine>::Scalar,
-}
-
-impl FastSerde for ProvingParams {
-  /// Initialize ProvingParams from an efficiently serializable data format.
-  fn from_bytes(bytes: &Vec<u8>) -> Result<Self, SerdeByteError> {
-    let mut cursor = Cursor::new(bytes);
-    Self::validate_header(&mut cursor, SerdeByteTypes::ProverParams, 3)?;
-
-    let aux_params =
-      Self::read_section_bytes(&mut cursor, 1).map(|bytes| AuxParams::from_bytes(&bytes))??;
-
-    let vk_digest_primary = Self::read_section_bytes(&mut cursor, 2)
-      .and_then(|bytes| bytes.try_into().map_err(|_| SerdeByteError::G1DecodeError))
-      .map(|bytes| <E1 as Engine>::Scalar::from_bytes(&bytes))?
-      .into_option()
-      .ok_or(SerdeByteError::G1DecodeError)?;
-
-    let vk_digest_secondary = Self::read_section_bytes(&mut cursor, 3)
-      .and_then(|bytes| bytes.try_into().map_err(|_| SerdeByteError::G2DecodeError))
-      .map(|bytes| <Dual<E1> as Engine>::Scalar::from_bytes(&bytes))?
-      .into_option()
-      .ok_or(SerdeByteError::G1DecodeError)?;
-
-    Ok(ProvingParams { aux_params, vk_digest_primary, vk_digest_secondary })
-  }
-
-  /// Convert ProvingParams to an efficient serialization.
-  fn to_bytes(&self) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(&fast_serde::MAGIC_NUMBER);
-    out.push(SerdeByteTypes::ProverParams as u8);
-    out.push(3); // num_sections
-
-    Self::write_section_bytes(&mut out, 1, &self.aux_params.to_bytes());
-    Self::write_section_bytes(&mut out, 2, &self.vk_digest_primary.to_bytes().to_vec());
-    Self::write_section_bytes(&mut out, 3, &self.vk_digest_secondary.to_bytes().to_vec());
-
-    return out;
-  }
-}
-
-impl ProvingParams {
-  /// Method used externally to initialize all the backend data needed to create a verifiable proof
-  /// with [`client_side_prover`] and `proofs` crate. Intended to be used to create these values
-  /// offline and then be loaded at or before proof creation or verification.
-  ///
-  /// # Arguments
-  /// - `aux_params`: the data that defines what types of supernova programs can be run, i.e.,
-  ///   specified by a list of circuit R1CS and max ROM length.
-  /// - `prover_key`: The key used for generating proofs, allows us to pin a specific verifier.
-  pub fn new(aux_params: AuxParams, prover_key: ProverKey) -> Result<ProvingParams, ProofError> {
-    Ok(ProvingParams {
-      aux_params,
-      vk_digest_primary: prover_key.pk_primary.vk_digest,
-      vk_digest_secondary: prover_key.pk_secondary.vk_digest,
-    })
-  }
-}
